# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Making knowledge workers more productive",
        "title": "General Task API",
        "termsOfService": "https://generaltask.com/terms-of-service",
        "contact": {
            "name": "Support",
            "email": "support@generaltask.com"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/link/{sourceID}/": {
            "get": {
                "description": "First step in oauth verification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Redirects to link callback for that service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "URL redirect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "service not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping/": {
            "get": {
                "description": "used to determine if server online",
                "tags": [
                    "utils"
                ],
                "summary": "Returns success",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}